Update version of yargs to address security issue
This addresses a security issue with `mem`, which is a dependency of
`os-locale`, which is a dependency of `yargs`, which is a dependency of
`elm-verify-examples`. Unfortunately addressing the issue requires us to
update `yargs` two major versions.

Luckily the major versions don't seem to introduce a whole bunch of
breaking changes and, as far as I can see, none relevant for
`elm-verify-examples`.

Yargs changelog:
https://github.com/yargs/yargs/blob/master/CHANGELOG.md

Breaking changes in version 13.x
- options with leading '+' or '0' now parse as strings
- dropping Node 6 which hits end of life in April 2019
- we now warn if the yargs stanza package.json is used.

Breaking changes in version 12.x:
- Options absent from argv (not set via CLI argument) are now absent from the parsed result object rather than being set with undefined
- drop Node 4 from testing matrix, such that we'll gradually start drifting away from supporting Node 4.
- yargs-parser does not populate 'false' when boolean flag is not passed
- tests that assert against help output will need to be updated