Fix cli for new version of yargs
Turns out the version of yargs we upgraded too did introduce breaking
changes relevant to `elm-verify-examples`. Specifically boolean options
now need to be explicitly marked as such, or using them without a value
will not work.

When options are marked as being boolean yargs automatically assumes
their absence to mean `false`, so we can remove the `default` clauses
for these.